api/certificate/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
jinja2==3.1.3
weasyprint==61.0
python-dotenv==1.0.1